+++ src/lib/server/mongodb.ts
@@ -1,0 +1,52 @@
+import mongoose from "mongoose";
+
+if (!process.env.NEXT_PUBLIC_MONGODB_URI) {
+  throw new Error("MONGODB_URI environment variable is not set");
+}
+
+const MONGODB_URI = process.env.NEXT_PUBLIC_MONGODB_URI!;
+
+// Define the type for the cached connection
+interface MongooseCache {
+  conn: typeof mongoose | null;
+  promise: Promise<typeof mongoose> | null;
+}
+
+// Declare the global mongoose property
+declare global {
+  var mongoose: MongooseCache | undefined;
+}
+
+// Initialize the cached connection
+let cached: MongooseCache = global.mongoose || { conn: null, promise: null };
+
+if (!global.mongoose) {
+  global.mongoose = cached;
+}
+
+async function connectDB() {
+  if (cached.conn) {
+    return cached.conn;
+  }
+
+  if (!cached.promise) {
+    const opts = {
+      bufferCommands: false,
+    };
+
+    cached.promise = mongoose.connect(MONGODB_URI, opts).then((mongoose) => {
+      return mongoose;
+    });
+  }
+
+  try {
+    cached.conn = await cached.promise;
+  } catch (e) {
+    cached.promise = null;
+    throw e;
+  }
+
+  return cached.conn;
+}
+
+export default connectDB;